FIX: reflect filter information correctly in blades
