+++ solid_attenuator/util.py
@@ -1,3 +1,4 @@
+import enum
 import functools
 import sys
 import typing
@@ -7,6 +8,65 @@ import caproto._log as caproto_log
 import caproto.threading
 
 _default_thread_context = None
+
+
+class State(enum.IntEnum):
+    """
+    State which matches that of the motion IOC.
+    """
+    # 'Moving' is also: "unknown" or "between states"
+    Moving = 0
+
+    # 'Out' is fixed at 1:
+    Out = 1
+
+    # And any "in" states follow:
+    In_01 = 2
+    In_02 = 3
+    In_03 = 4
+    In_04 = 5
+    In_05 = 6
+    In_06 = 7
+    In_07 = 8
+    In_08 = 9
+    In_09 = 10
+
+    @property
+    def filter_index(self) -> typing.Optional[int]:
+        """The one-based filter index, if inserted."""
+        if not self.is_inserted:
+            return None
+        return self.value - 1
+
+    @property
+    def is_inserted(self) -> bool:
+        """Is a filter inserted?"""
+        return self not in {State.Moving, State.Out}
+
+    @property
+    def is_moving(self) -> bool:
+        """Is the blade moving?"""
+        return self == State.Moving
+
+    @classmethod
+    def from_filter_index(self, idx: typing.Optional[int]) -> 'State':
+        """Get a State from a filter index (where filter 1 is 1)."""
+        return {
+            None: State.Out,
+            0: State.Out,
+            1: State.In_01,
+            2: State.In_02,
+            3: State.In_03,
+            4: State.In_04,
+            5: State.In_05,
+            6: State.In_06,
+            7: State.In_07,
+            8: State.In_08,
+            9: State.In_09,
+        }[idx]
+
+    def __repr__(self):
+        return self.name
 
 
 def get_default_thread_context():
